Add files via upload
Does not work on jupyter/spyder due to weird internal issues with how multiprocessing works on ipython consoles. To run properly, you have to run the pure python script in the command window. There will be "variable not defined"/"index out of range" errors due to much of our code not being inside functions - for now, parallelization is on the backburner.

Better news - In the process I did simplify the redundant parts of our code into functions. Not yet implemented in the main Explore_RCIP file, but worth looking at.